py/requirements.txt
rich
python-dotenv
requests
parse
advent-of-code-ocr
beautifulsoup4
